Poster demo
I divided the basic layout and fonts as well as color matching. We had better choose the color matching at first, so that it is convenient to change the r code later to generate the corresponding image. If you have a better idea, please tell me on whatsapp or WeChat.